CURB-2777 Moved pipeline validation error helper from registration to account folder
- renamed helper function name to a neutral one
- fixed existing unit tests
- added unit test for attribute conversion
