+++ libraries/engage/registration/providers/GuestRegistrationProvider.actions.js
@@ -1,8 +1,10 @@
 import { EVALIDATION } from '@shopgate/engage/core';
 import { updateCheckoutOrder } from '@shopgate/engage/checkout/actions';
-import { extractAttributes } from '@shopgate/engage/account/helper/form';
+import {
+  extractAttributes,
+  convertPipelineValidationErrors,
+} from '@shopgate/engage/account/helper/form';
 import { getMerchantCustomerAttributes } from '@shopgate/engage/core/selectors/merchantSettings';
-import { convertSubmitRegistrationValidationErrors } from '../helpers';
 
 /**
  * Submits guest registration form data.
@@ -70,7 +72,7 @@ export const submitGuestRegistration = ({
     }
   }
 
-  const converted = convertSubmitRegistrationValidationErrors(errors, attributes);
+  const converted = convertPipelineValidationErrors(errors, attributes);
 
   if (converted?.validation && Object.keys(converted.validation).length > 0) {
     const sequenceErrors = converted?.validation?.addressSequences;
